refactor: rearrange images in each folder
