+++ skirunmap/src/components/LikeDislike/images/clicked-like-arrow.png
