+++ skirunmap/src/pages/Member/index.tsx
@@ -15,20 +15,20 @@ import { useEffect, useState } from 'react'
 import { Link, useParams } from 'react-router-dom'
 import { db } from '../../auth/Firebase'
 import RouteCard from '../../components/RouteCard'
+import Bookmark from '../../images/bookmark.png'
+import ViewIcon from '../../images/eye.png'
+import SkiIcon from '../../images/skiing-icon.png'
+import SnowboardIcon from '../../images/snowboarder-icon.png'
 import { useRouteCardStore } from '../../store/useRouteCard'
 import { User, useUserStore } from '../../store/useUser'
-import AboutMeIcon from './about-me-icon.png'
-import Bookmark from './bookmark.png'
-import CalenderIcon from './calendar-icon.png'
-import CountryIcon from './country-icon.png'
-import EditIcon from './edit.png'
-import ViewIcon from './eye.png'
-import FollowerIcon from './follower-icon.png'
-import FriendIcon from './friend-icon.png'
-import GenderIcon from './gender-icon.png'
-import RouteIcon from './route-icon.png'
-import SkiIcon from './skiing-icon.png'
-import SnowboardIcon from './snowboarder-icon.png'
+import AboutMeIcon from './images/about-me-icon.png'
+import CalenderIcon from './images/calendar-icon.png'
+import CountryIcon from './images/country-icon.png'
+import EditIcon from './images/edit.png'
+import FollowerIcon from './images/follower-icon.png'
+import FriendIcon from './images/friend-icon.png'
+import GenderIcon from './images/gender-icon.png'
+import RouteIcon from './images/route-icon.png'
 
 interface RouteDocsInList {
   listName: string
